learning/cs50/bgremover/requirements.txt
Pillow==10.1.0
rembg==2.0.55
numpy==1.26.0
pytest==7.4.2
